chapter4/pyproject.toml
[project]
name = "chapter4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3==1.40.21",
    "ipykernel>=7.1.0",
    "langchain==0.3.27",
    "langchain-aws==0.2.31",
    "langchain-community==0.3.29",
    "langchain-mcp-adapters==0.1.9",
    "langchain-tavily==0.2.11",
    "langgraph==0.6.6",
    "pydantic==2.11.7",
    "python-dotenv==1.1.1",
    "rich>=14.2.0",
    "streamlit==1.49.1",
    "streamlit-jupyter>=0.3.1",
    "uv==0.8.14",
]
